% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-cli.R
\name{opensimCLIAvailable}
\alias{opensimCLIAvailable}
\title{Check Whether OpenSim CLI Is Available}
\usage{
opensimCLIAvailable(cli = NULL)
}
\arguments{
\item{cli}{Optional command or absolute path.}
}
\value{
\code{TRUE} if CLI command is resolvable, otherwise \code{FALSE}.
}
\description{
Check Whether OpenSim CLI Is Available
}
\examples{
opensimCLIAvailable()
}
