% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-cli.R
\name{opensimCLIPath}
\alias{opensimCLIPath}
\title{Resolve OpenSim CLI Command Path}
\usage{
opensimCLIPath(cli = NULL)
}
\arguments{
\item{cli}{Optional command or absolute path. Defaults to \code{OPENSIM_CLI}
environment variable, then \code{"opensim-cmd"}.}
}
\value{
A character scalar path to executable command.
}
\description{
Resolves the command path used to invoke OpenSim CLI.
}
\examples{
\dontrun{
opensimCLIPath()
opensimCLIPath("/usr/local/bin/opensim-cmd")
}
}
