% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-native.R
\name{opensimFinalizeConnections}
\alias{opensimFinalizeConnections}
\title{Finalize OpenSim Model Connections}
\usage{
opensimFinalizeConnections(model)
}
\arguments{
\item{model}{A \code{PhysioOpenSimModel} external pointer.}
}
\value{
The input \code{model} handle (invisibly).
}
\description{
Finalize OpenSim Model Connections
}
\examples{
\dontrun{
model <- opensimLoadModel("gait2392.osim")
opensimFinalizeConnections(model)
}
}
