% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-native.R
\name{opensimLoadModel}
\alias{opensimLoadModel}
\title{Load an OpenSim Model as a Native External Pointer}
\usage{
opensimLoadModel(path)
}
\arguments{
\item{path}{Path to an OpenSim model file (\code{.osim}).}
}
\value{
A \code{PhysioOpenSimModel} external pointer handle.
}
\description{
Load an OpenSim Model as a Native External Pointer
}
\examples{
\dontrun{
model <- opensimLoadModel("gait2392.osim")
opensimModelName(model)
}
}
