% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-native.R
\name{opensimModelComponents}
\alias{opensimModelComponents}
\title{List Core OpenSim Model Component Names}
\usage{
opensimModelComponents(model)
}
\arguments{
\item{model}{A \code{PhysioOpenSimModel} external pointer.}
}
\value{
Named list with vectors for body, joint, marker, muscle, and
coordinate names.
}
\description{
List Core OpenSim Model Component Names
}
\examples{
\dontrun{
model <- opensimLoadModel("gait2392.osim")
comps <- opensimModelComponents(model)
comps$bodies
comps$muscles
}
}
