% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-native.R
\name{opensimModelInitialize}
\alias{opensimModelInitialize}
\title{Initialize OpenSim Model System State}
\usage{
opensimModelInitialize(model)
}
\arguments{
\item{model}{A \code{PhysioOpenSimModel} external pointer.}
}
\value{
The input \code{model} handle (invisibly).
}
\description{
Initialize OpenSim Model System State
}
\examples{
\dontrun{
model <- opensimLoadModel("gait2392.osim")
opensimModelInitialize(model)
opensimModelIsInitialized(model)  # TRUE
}
}
