% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-native.R
\name{opensimModelIsInitialized}
\alias{opensimModelIsInitialized}
\title{Check Whether an OpenSim Model Handle Is Initialized}
\usage{
opensimModelIsInitialized(model)
}
\arguments{
\item{model}{A \code{PhysioOpenSimModel} external pointer.}
}
\value{
\code{TRUE} when model state has been initialized by OpenSim.
}
\description{
Check Whether an OpenSim Model Handle Is Initialized
}
\examples{
\dontrun{
model <- opensimLoadModel("gait2392.osim")
opensimModelIsInitialized(model)
}
}
