% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-native.R
\name{opensimModelName}
\alias{opensimModelName}
\title{Get OpenSim Model Name}
\usage{
opensimModelName(model)
}
\arguments{
\item{model}{A \code{PhysioOpenSimModel} external pointer.}
}
\value{
Character scalar model name.
}
\description{
Get OpenSim Model Name
}
\examples{
\dontrun{
model <- opensimLoadModel("gait2392.osim")
opensimModelName(model)
}
}
