% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-native.R
\name{opensimSaveModel}
\alias{opensimSaveModel}
\title{Save OpenSim Model to File}
\usage{
opensimSaveModel(model, output_file)
}
\arguments{
\item{model}{A \code{PhysioOpenSimModel} external pointer.}

\item{output_file}{Output path for model XML.}
}
\value{
Normalized output path.
}
\description{
Save OpenSim Model to File
}
\examples{
\dontrun{
model <- opensimLoadModel("gait2392.osim")
opensimSetModelName(model, "modified")
opensimSaveModel(model, "gait2392_modified.osim")
}
}
