% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensim-native.R
\name{opensimSetModelName}
\alias{opensimSetModelName}
\title{Set OpenSim Model Name}
\usage{
opensimSetModelName(model, name)
}
\arguments{
\item{model}{A \code{PhysioOpenSimModel} external pointer.}

\item{name}{New model name.}
}
\value{
The input \code{model} handle (invisibly).
}
\description{
Set OpenSim Model Name
}
\examples{
\dontrun{
model <- opensimLoadModel("gait2392.osim")
opensimSetModelName(model, "my_model")
opensimModelName(model)  # "my_model"
}
}
